backend/app/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
openai
pydantic
aiofiles
requests
httpx